prevent admin from delete his account

+++ dashboard/src/app/modules/admin/user/user.component.ts
@@ -2,6 +2,8 @@ import { Component, OnInit, ViewEncapsulation } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
 import { MatTableDataSource } from '@angular/material/table';
 import { Router } from '@angular/router';
+import { fuseAnimations } from '@fuse/animations';
+import { FuseAlertType } from '@fuse/components/alert';
 import { UserService } from 'app/core/user/user.service';
 import { ConfirmDialogComponent } from 'app/shared/dialog/confirm-dialog.component';
 
@@ -9,6 +11,7 @@ import { ConfirmDialogComponent } from 'app/shared/dialog/confirm-dialog.compone
     selector: 'app-user',
     templateUrl: './user.component.html',
     encapsulation: ViewEncapsulation.None,
+    animations: fuseAnimations,
 })
 export class UserComponent implements OnInit {
     usersDataSource: MatTableDataSource<any> = new MatTableDataSource();
@@ -23,7 +26,11 @@ export class UserComponent implements OnInit {
         1: 'Client',
         99: 'Admin',
     };
-
+    alert: { type: FuseAlertType; message: string } = {
+        type: 'success',
+        message: '',
+    };
+    showAlert: boolean = false;
     ngOnInit(): void {
         this.userService.getUsers().subscribe(
             (data: any) => {
@@ -54,6 +61,16 @@ export class UserComponent implements OnInit {
 
         dialogRef.afterClosed().subscribe((result) => {
             if (result === true) {
+                this.showAlert = false
+                if(this.userService.getLoggedInUser()._id == userId){
+                    this.alert = {
+                        type: 'error',
+                        message:
+                            "You can't delete your owen account",
+                    };
+                    this.showAlert = true
+                    return
+                }
                 this.userService.deleteUser(userId).subscribe((res) => {
                     this.usersDataSource.data =
                         this.usersDataSource.data.filter((u) => {
